Clean discount related base mutations
