+++ saleor/graphql/discount/tests/mutations/test_sale_create.py
@@ -46,7 +46,7 @@ SALE_CREATE_MUTATION = """
 def test_create_sale(
     created_webhook_mock,
     sale_toggle_mock,
-    update_products_discounted_prices_of_sale_task_mock,
+    update_products_discounted_prices_of_promotion_task_mock,
     staff_api_client,
     permission_manage_discounts,
     product_list,
@@ -100,7 +100,9 @@ def test_create_sale(
     )
     created_webhook_mock.assert_called_once_with(sale, current_catalogue)
     sale_toggle_mock.assert_called_once_with(sale, current_catalogue)
-    update_products_discounted_prices_of_sale_task_mock.assert_called_once_with(sale.id)
+    update_products_discounted_prices_of_promotion_task_mock.assert_called_once_with(
+        sale.id
+    )
 
 
 @freeze_time("2020-03-18 12:00:00")
@@ -110,7 +112,7 @@ def test_create_sale(
 def test_create_sale_only_start_date(
     created_webhook_mock,
     sale_toggle_mock,
-    update_products_discounted_prices_of_sale_task_mock,
+    update_products_discounted_prices_of_promotion_task_mock,
     staff_api_client,
     permission_manage_discounts,
     product_list,
@@ -152,12 +154,14 @@ def test_create_sale_only_start_date(
     )
     created_webhook_mock.assert_called_once_with(sale, current_catalogue)
     sale_toggle_mock.assert_called_once_with(sale, current_catalogue)
-    update_products_discounted_prices_of_sale_task_mock.assert_called_once_with(sale.id)
+    update_products_discounted_prices_of_promotion_task_mock.assert_called_once_with(
+        sale.id
+    )
 
 
 @patch("saleor.product.tasks.update_products_discounted_prices_of_promotion_task.delay")
 def test_create_sale_with_end_date_before_startdate(
-    update_products_discounted_prices_of_sale_task_mock,
+    update_products_discounted_prices_of_promotion_task_mock,
     staff_api_client,
     permission_manage_discounts,
 ):
@@ -186,7 +190,7 @@ def test_create_sale_with_end_date_before_startdate(
     assert len(errors) == 1
     assert errors[0]["field"] == "endDate"
     assert errors[0]["code"] == DiscountErrorCode.INVALID.name
-    update_products_discounted_prices_of_sale_task_mock.assert_not_called()
+    update_products_discounted_prices_of_promotion_task_mock.assert_not_called()
 
 
 @patch("saleor.product.tasks.update_products_discounted_prices_of_promotion_task.delay")
@@ -195,7 +199,7 @@ def test_create_sale_with_end_date_before_startdate(
 def test_create_sale_start_date_and_end_date_before_current_date(
     created_webhook_mock,
     sale_toggle_mock,
-    update_products_discounted_prices_of_sale_task_mock,
+    update_products_discounted_prices_of_promotion_task_mock,
     staff_api_client,
     permission_manage_discounts,
     product_list,
@@ -241,7 +245,9 @@ def test_create_sale_start_date_and_end_date_before_current_date(
     )
     created_webhook_mock.assert_called_once_with(sale, current_catalogue)
     sale_toggle_mock.assert_not_called()
-    update_products_discounted_prices_of_sale_task_mock.assert_called_once_with(sale.id)
+    update_products_discounted_prices_of_promotion_task_mock.assert_called_once_with(
+        sale.id
+    )
 
 
 @patch("saleor.product.tasks.update_products_discounted_prices_of_promotion_task.delay")
@@ -250,7 +256,7 @@ def test_create_sale_start_date_and_end_date_before_current_date(
 def test_create_sale_start_date_and_end_date_after_current_date(
     created_webhook_mock,
     sale_toggle_mock,
-    update_products_discounted_prices_of_sale_task_mock,
+    update_products_discounted_prices_of_promotion_task_mock,
     staff_api_client,
     permission_manage_discounts,
     product_list,
@@ -296,7 +302,9 @@ def test_create_sale_start_date_and_end_date_after_current_date(
     )
     created_webhook_mock.assert_called_once_with(sale, current_catalogue)
     sale_toggle_mock.assert_not_called()
-    update_products_discounted_prices_of_sale_task_mock.assert_called_once_with(sale.id)
+    update_products_discounted_prices_of_promotion_task_mock.assert_called_once_with(
+        sale.id
+    )
 
 
 @freeze_time("2020-03-18 12:00:00")
